Add retry logic to cmd, use for GitRemoteResolver
I added an optional param to cmd to specify retry logic. If specified,
the call to throttler.enqueue will be wrapped in retry logic (from the
promise-retry package).

I use this to add retries to GitRemoteResolver#refs which calls `git
remote-ls --tags` on github.com. Unfortunately, there appears to be
pretty strict rate limiting on calling this and without any retries, the
entire bower operation will fail.

Even without the rate limiting, it seems to me to be a good idea to be
able to retry idempotent operations.